WorkingDay Ulices

+++ src/main/java/ar/com/academy/mfs/repository/ZoneRepository.java
@@ -6,5 +6,5 @@ import ar.com.academy.mfs.model.Zone;
 
 public interface ZoneRepository extends JpaRepository<Zone, Integer> {
 	Zone findByZoneName(String zoneName);
-	Zone findByZoneId(long zone_id);
+	Zone findByZoneId(int zone_id);
 }
